+++ tests/e2e/modules/alert_spec.js
@@ -1,10 +1,5 @@
 const helpers = require("../global-setup");
 
-const describe = global.describe;
-const it = global.it;
-const beforeEach = global.beforeEach;
-const afterEach = global.afterEach;
-
 describe("Alert module", function () {
 	helpers.setupTimeout(this);
 
@@ -25,7 +20,7 @@ describe("Alert module", function () {
 	});
 
 	describe("Default configuration", function () {
-		before(function () {
+		beforeAll(function () {
 			// Set config sample for use in test
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/alert/default.js";
 		});