+++ tests/e2e/modules/weather_spec.js
@@ -1,4 +1,3 @@
-const expect = require("chai").expect;
 const fs = require("fs");
 const moment = require("moment");
 const path = require("path");
@@ -31,7 +30,7 @@ describe("Weather module", function () {
 	async function getText(element, result) {
 		const elem = await getElement(element);
 		return await elem.getText(element).then(function (text) {
-			expect(text.trim()).to.equal(result);
+			expect(text.trim()).toBe(result);
 		});
 	}
 
@@ -42,12 +41,12 @@ describe("Weather module", function () {
 	describe("Current weather", function () {
 		let template;
 
-		before(function () {
+		beforeAll(function () {
 			template = fs.readFileSync(path.join(__dirname, "..", "..", "..", "modules", "default", "weather", "current.njk"), "utf8");
 		});
 
 		describe("Default configuration", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/currentweather_default.js";
 			});
 
@@ -94,7 +93,7 @@ describe("Weather module", function () {
 		});
 
 		describe("Compliments Integration", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/currentweather_compliments.js";
 			});
 
@@ -107,7 +106,7 @@ describe("Weather module", function () {
 		});
 
 		describe("Configuration Options", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/currentweather_options.js";
 			});
 
@@ -124,7 +123,7 @@ describe("Weather module", function () {
 
 				const elem = await getElement(".weather .normal.medium sup i.fa-long-arrow-up");
 				return elem.getHTML(".weather .normal.medium sup i.fa-long-arrow-up").then(function (text) {
-					expect(text).to.include("transform:rotate(250deg);");
+					expect(text).toContain("transform:rotate(250deg);");
 				});
 			});
 
@@ -144,7 +143,7 @@ describe("Weather module", function () {
 		});
 
 		describe("Current weather units", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/currentweather_units.js";
 			});
 
@@ -185,12 +184,12 @@ describe("Weather module", function () {
 	describe("Weather Forecast", function () {
 		let template;
 
-		before(function () {
+		beforeAll(function () {
 			template = fs.readFileSync(path.join(__dirname, "..", "..", "..", "modules", "default", "weather", "forecast.njk"), "utf8");
 		});
 
 		describe("Default configuration", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/forecastweather_default.js";
 			});
 
@@ -248,13 +247,13 @@ describe("Weather module", function () {
 
 				for (const [index, opacity] of opacities.entries()) {
 					const html = await elem.getHTML(`.weather table.small tr:nth-child(${index + 1})`);
-					expect(html).to.includes(`<tr style="opacity: ${opacity};">`);
+					expect(html).toContain(`<tr style="opacity: ${opacity};">`);
 				}
 			});
 		});
 
 		describe("Configuration Options", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/forecastweather_options.js";
 			});
 
@@ -271,12 +270,12 @@ describe("Weather module", function () {
 
 				const rows = await app.client.$$(".weather table.myTableClass tr.colored");
 
-				expect(rows.length).to.be.equal(5);
+				expect(rows.length).toBe(5);
 			});
 		});
 
 		describe("Forecast weather units", function () {
-			before(function () {
+			beforeAll(function () {
 				process.env.MM_CONFIG_FILE = "tests/configs/modules/weather/forecastweather_units.js";
 			});