+++ tests/e2e/port_config.js
@@ -1,11 +1,5 @@
 const helpers = require("./global-setup");
 const fetch = require("node-fetch");
-const expect = require("chai").expect;
-
-const describe = global.describe;
-const it = global.it;
-const beforeEach = global.beforeEach;
-const afterEach = global.afterEach;
 
 describe("port directive configuration", function () {
 	helpers.setupTimeout(this);
@@ -27,33 +21,33 @@ describe("port directive configuration", function () {
 	});
 
 	describe("Set port 8090", function () {
-		before(function () {
+		beforeAll(function () {
 			// Set config sample for use in this test
 			process.env.MM_CONFIG_FILE = "tests/configs/port_8090.js";
 		});
 
 		it("should return 200", function (done) {
 			fetch("http://localhost:8090").then((res) => {
-				expect(res.status).to.equal(200);
+				expect(res.status).toBe(200);
 				done();
 			});
 		});
 	});
 
 	describe("Set port 8100 on environment variable MM_PORT", function () {
-		before(function () {
+		beforeAll(function () {
 			process.env.MM_PORT = 8100;
 			// Set config sample for use in this test
 			process.env.MM_CONFIG_FILE = "tests/configs/port_8090.js";
 		});
 
-		after(function () {
+		afterAll(function () {
 			delete process.env.MM_PORT;
 		});
 
 		it("should return 200", function (done) {
 			fetch("http://localhost:8100").then((res) => {
-				expect(res.status).to.equal(200);
+				expect(res.status).toBe(200);
 				done();
 			});
 		});