+++ tests/e2e/translations_spec.js
@@ -1,8 +1,5 @@
 const fs = require("fs");
 const path = require("path");
-const chai = require("chai");
-const expect = chai.expect;
-const mlog = require("mocha-logger");
 const translations = require("../../translations/translations.js");
 const helmet = require("helmet");
 const { JSDOM } = require("jsdom");
@@ -12,7 +9,7 @@ const sinon = require("sinon");
 describe("Translations", function () {
 	let server;
 
-	before(function () {
+	beforeAll(function () {
 		const app = express();
 		app.use(helmet());
 		app.use(function (req, res, next) {
@@ -24,14 +21,14 @@ describe("Translations", function () {
 		server = app.listen(3000);
 	});
 
-	after(function () {
+	afterAll(function () {
 		server.close();
 	});
 
 	it("should have a translation file in the specified path", function () {
 		for (let language in translations) {
 			const file = fs.statSync(translations[language]);
-			expect(file.isFile()).to.be.equal(true);
+			expect(file.isFile()).toBe(true);
 		}
 	});
 
@@ -59,9 +56,9 @@ describe("Translations", function () {
 				const loaded = sinon.stub();
 				MMM.loadTranslations(loaded);
 
-				expect(loaded.callCount).to.equal(1);
-				expect(Translator.load.args.length).to.equal(1);
-				expect(Translator.load.calledWith(MMM, "translations/en.json", false, sinon.match.func)).to.be.true;
+				expect(loaded.callCount).toBe(1);
+				expect(Translator.load.args.length).toBe(1);
+				expect(Translator.load.calledWith(MMM, "translations/en.json", false, sinon.match.func)).toBe(true);
 
 				done();
 			};
@@ -78,10 +75,10 @@ describe("Translations", function () {
 				const loaded = sinon.stub();
 				MMM.loadTranslations(loaded);
 
-				expect(loaded.callCount).to.equal(1);
-				expect(Translator.load.args.length).to.equal(2);
-				expect(Translator.load.calledWith(MMM, "translations/de.json", false, sinon.match.func)).to.be.true;
-				expect(Translator.load.calledWith(MMM, "translations/en.json", true, sinon.match.func)).to.be.true;
+				expect(loaded.callCount).toBe(1);
+				expect(Translator.load.args.length).toBe(2);
+				expect(Translator.load.calledWith(MMM, "translations/de.json", false, sinon.match.func)).toBe(true);
+				expect(Translator.load.calledWith(MMM, "translations/en.json", true, sinon.match.func)).toBe(true);
 
 				done();
 			};
@@ -99,9 +96,9 @@ describe("Translations", function () {
 				const loaded = sinon.stub();
 				MMM.loadTranslations(loaded);
 
-				expect(loaded.callCount).to.equal(1);
-				expect(Translator.load.args.length).to.equal(1);
-				expect(Translator.load.calledWith(MMM, "translations/en.json", true, sinon.match.func)).to.be.true;
+				expect(loaded.callCount).toBe(1);
+				expect(Translator.load.args.length).toBe(1);
+				expect(Translator.load.calledWith(MMM, "translations/en.json", true, sinon.match.func)).toBe(true);
 
 				done();
 			};
@@ -118,8 +115,8 @@ describe("Translations", function () {
 				const loaded = sinon.stub();
 				MMM.loadTranslations(loaded);
 
-				expect(loaded.callCount).to.equal(1);
-				expect(Translator.load.callCount).to.equal(0);
+				expect(loaded.callCount).toBe(1);
+				expect(Translator.load.callCount).toBe(0);
 
 				done();
 			};
@@ -145,8 +142,8 @@ describe("Translations", function () {
 					const { Translator } = dom.window;
 
 					Translator.load(mmm, translations[language], false, function () {
-						expect(Translator.translations[mmm.name]).to.be.an("object");
-						expect(Object.keys(Translator.translations[mmm.name]).length).to.be.at.least(1);
+						expect(typeof Translator.translations[mmm.name]).toBe("object");
+						expect(Object.keys(Translator.translations[mmm.name]).length).toBeGreaterThanOrEqual(1);
 						done();
 					});
 				};
@@ -156,8 +153,9 @@ describe("Translations", function () {
 
 	describe("Same keys", function () {
 		let base;
+		let missing = [];
 
-		before(function (done) {
+		beforeAll(function (done) {
 			const dom = new JSDOM(
 				`<script>var translations = ${JSON.stringify(translations)}; var Log = {log: function(){}};</script>\
 					<script src="file://${path.join(__dirname, "..", "..", "js", "translator.js")}">`,
@@ -173,6 +171,10 @@ describe("Translations", function () {
 			};
 		});
 
+		afterAll(function () {
+			console.log(missing);
+		});
+
 		for (let language in translations) {
 			if (language === "en") {
 				continue;
@@ -181,7 +183,7 @@ describe("Translations", function () {
 			describe(`Translation keys of ${language}`, function () {
 				let keys;
 
-				before(function (done) {
+				beforeAll(function (done) {
 					const dom = new JSDOM(
 						`<script>var translations = ${JSON.stringify(translations)}; var Log = {log: function(){}};</script>\
 					<script src="file://${path.join(__dirname, "..", "..", "js", "translator.js")}">`,
@@ -199,22 +201,21 @@ describe("Translations", function () {
 
 				it(`${language} keys should be in base`, function () {
 					keys.forEach(function (key) {
-						expect(base.indexOf(key)).to.be.at.least(0);
+						expect(base.indexOf(key)).toBeGreaterThanOrEqual(0);
 					});
 				});
 
 				it(`${language} should contain all base keys`, function () {
 					// TODO: when all translations are fixed, use
-					// expect(keys).to.deep.equal(base);
+					// expect(keys).toEqual(base);
 					// instead of the try-catch-block
 
 					try {
-						expect(keys).to.deep.equal(base);
+						expect(keys).toEqual(base);
 					} catch (e) {
-						if (e instanceof chai.AssertionError) {
+						if (e.message.match(/expect.*toEqual/)) {
 							const diff = base.filter((key) => !keys.includes(key));
-							mlog.pending(`Missing Translations for language ${language}: ${diff}`);
-							this.skip();
+							missing.push(`Missing Translations for language ${language}: ${diff}`);
 						} else {
 							throw e;
 						}