+++ tests/unit/functions/currentweather_spec.js
@@ -1,6 +1,4 @@
 /* eslint no-multi-spaces: 0 */
-const expect = require("chai").expect;
-
 describe("Functions module currentweather", function () {
 	// Fake for use by currentweather.js
 	Module = {};
@@ -10,14 +8,14 @@ describe("Functions module currentweather", function () {
 		Module.definitions[name] = moduleDefinition;
 	};
 
-	before(function () {
+	beforeAll(function () {
 		require("../../../modules/default/currentweather/currentweather.js");
 		Module.definitions.currentweather.config = {};
 	});
 
 	describe("roundValue", function () {
 		describe("this.config.roundTemp is true", function () {
-			before(function () {
+			beforeAll(function () {
 				Module.definitions.currentweather.config.roundTemp = true;
 			});
 
@@ -35,13 +33,13 @@ describe("Functions module currentweather", function () {
 
 			values.forEach((value) => {
 				it(`for ${value[0]} should be return ${value[1]}`, function () {
-					expect(Module.definitions.currentweather.roundValue(value[0])).to.equal(value[1]);
+					expect(Module.definitions.currentweather.roundValue(value[0])).toBe(value[1]);
 				});
 			});
 		});
 
 		describe("this.config.roundTemp is false", function () {
-			before(function () {
+			beforeAll(function () {
 				Module.definitions.currentweather.config.roundTemp = false;
 			});
 
@@ -60,7 +58,7 @@ describe("Functions module currentweather", function () {
 
 			values.forEach((value) => {
 				it(`for ${value[0]} should be return ${value[1]}`, function () {
-					expect(Module.definitions.currentweather.roundValue(value[0])).to.equal(value[1]);
+					expect(Module.definitions.currentweather.roundValue(value[0])).toBe(value[1]);
 				});
 			});
 		});