+++ tests/unit/global_vars/defaults_modules_spec.js
@@ -1,39 +1,31 @@
 const fs = require("fs");
 const path = require("path");
-const expect = require("chai").expect;
 const vm = require("vm");
 
 const basedir = path.join(__dirname, "../../..");
 
-before(function () {
+beforeAll(function () {
 	const fileName = "js/app.js";
 	const filePath = path.join(basedir, fileName);
 	const code = fs.readFileSync(filePath);
 
-	this.sandbox = {
+	sandbox = {
 		module: {},
 		__dirname: path.dirname(filePath),
 		global: {},
-		console: {
-			log: function () {
-				/*console.log("console.log(", arguments, ")");*/
-			}
-		},
 		process: {
-			on: function () {
-				/*console.log("process.on called with: ", arguments);*/
-			},
+			on: function () {},
 			env: {}
 		}
 	};
 
-	this.sandbox.require = function (filename) {
+	sandbox.require = function (filename) {
 		// This modifies the global slightly,
 		// but supplies vm with essential code
 		return require(filename);
 	};
 
-	vm.runInNewContext(code, this.sandbox, fileName);
+	vm.runInNewContext(code, sandbox, fileName);
 });
 
 describe("Default modules set in modules/default/defaultmodules.js", function () {
@@ -41,7 +33,7 @@ describe("Default modules set in modules/default/defaultmodules.js", function ()
 
 	for (const defaultModule of expectedDefaultModules) {
 		it(`contains a folder for modules/default/${defaultModule}"`, function () {
-			expect(fs.existsSync(path.join(this.sandbox.global.root_path, "modules/default", defaultModule))).to.equal(true);
+			expect(fs.existsSync(path.join(sandbox.global.root_path, "modules/default", defaultModule))).toBe(true);
 		});
 	}
 });