+++ tests/unit/global_vars/root_path_spec.js
@@ -1,43 +1,31 @@
 const fs = require("fs");
 const path = require("path");
-const expect = require("chai").expect;
 const vm = require("vm");
 
-before(function () {
+beforeAll(function () {
 	const basedir = path.join(__dirname, "../../..");
 
 	const fileName = "js/app.js";
 	const filePath = path.join(basedir, fileName);
 	const code = fs.readFileSync(filePath);
 
-	this.sandbox = {
+	sandbox = {
 		module: {},
 		__dirname: path.dirname(filePath),
 		global: {},
-		console: {
-			log: function () {
-				/*console.log("console.log(", arguments, ")");*/
-			}
-		},
 		process: {
-			on: function () {
-				/*console.log("process.on called with: ", arguments);*/
-			},
+			on: function () {},
 			env: {}
 		}
 	};
 
-	this.sandbox.require = function (filename) {
+	sandbox.require = function (filename) {
 		// This modifies the global slightly,
 		// but supplies vm with essential code
 		return require(filename);
 	};
 
-	vm.runInNewContext(code, this.sandbox, fileName);
-});
-
-after(function () {
-	//console.log(global);
+	vm.runInNewContext(code, sandbox, fileName);
 });
 
 describe("'global.root_path' set in js/app.js", function () {
@@ -45,20 +33,20 @@ describe("'global.root_path' set in js/app.js", function () {
 
 	expectedSubPaths.forEach((subpath) => {
 		it(`contains a file/folder "${subpath}"`, function () {
-			expect(fs.existsSync(path.join(this.sandbox.global.root_path, subpath))).to.equal(true);
+			expect(fs.existsSync(path.join(sandbox.global.root_path, subpath))).toBe(true);
 		});
 	});
 
 	it("should not modify global.root_path for testing", function () {
-		expect(global.root_path).to.equal(undefined);
+		expect(global.root_path).toBe(undefined);
 	});
 
 	it("should not modify global.version for testing", function () {
-		expect(global.version).to.equal(undefined);
+		expect(global.version).toBe(undefined);
 	});
 
 	it("should expect the global.version equals package.json file", function () {
 		const versionPackage = JSON.parse(fs.readFileSync("package.json", "utf8")).version;
-		expect(this.sandbox.global.version).to.equal(versionPackage);
+		expect(sandbox.global.version).toBe(versionPackage);
 	});
 });