+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Special thanks to the following contributors: @B1gG, @codac, @ezeholz, @khassel,
 - Added test case for recurring calendar events
 - Added new Environment Canada provider for default WEATHER module (weather data for Canadian locations only)
 - Added list view for newsfeed module.
+- Added dev dependency jest, switching from mocha to jest
 
 ### Updated
 
@@ -36,6 +37,8 @@ Special thanks to the following contributors: @B1gG, @codac, @ezeholz, @khassel,
 - Update documentation
 
 ### Removed
+
+- switching from mocha to jest so removed following dev dependencies: chai, chai-as-promised, mocha, mocha-each, mocha-logger
 
 ### Fixed